moving tests to lib folder #14 #15 #16
