+++ test/lib/repositories/service_type_repository/firebase/firebase_service_type_repository_test.mocks.dart
